requirements.txt
requests>=2.25.0
openai>=1.3.0
python-dotenv>=1.0.0
rich>=13.5.0
typer>=0.9.0
prompt_toolkit>=3.0.0